inference-docker/requirements.txt
opencv-python~=4.11.0.86
openslide-python~=1.4.1
tqdm~=4.67.1
ultralytics~=8.4.14
